// src/types/carrier/transport.ts
export interface RequestSnapshot {
  id: number;
  status: number;
  weight: number | null;
  arrival_to: number;
  trans_type: number;
  operator_id: string;
  total_width: number | null;
  total_height: number | null;
  total_length: number | null;
  operator_code: string | null;
  departure_from: number;
  price_per_unit: number;
  trailer_number: number;
  trans_order_id: string | null;
  transport_code: string;
  transport_name: string;
  operator_name: string;
  collection_date: string;
  cns_line_item_id: number | null;
  transport_plan_id: number | null;
  vehicle_condition: string | null;
  collection_time_to: string;
  outer_package_code: number;
  trailer_number_rest: number;
  collection_time_from: string;
  req_cns_line_item_id: number | null;
  special_instructions: string | null;
  temperature_range: string[] | null;
}
// ...
export interface PackageCNSLine {
  id: number;
  trans_type: number;
  operator_id: string;
  operator_name: string;
  cns_line_item_id: string;
  req_cns_line_item_id: string;
  trans_plan_id: string;
  collection_date: string;
  collection_time_from: string;
  collection_time_to: string;
  outer_package_code: number;
  price_per_unit: number;
  status: number;
  total_height: number;
  total_length: number;
  total_width: number;
  trailer_number: number;
  trailer_number_rest: number;
  weight: number;
  arrival_to: number;
  departure_from: number;
  transport_code: string;
  transport_name: string;
  special_instructions: string;
  vehicle_condition: string | null;
  cnsg_prty_name_txt: string;
  temperature_range: string[] | null;
  departure_time_max: string | null;
  departure_time_min: string | null;
  vehicle_avb_resource_id: number;
  cut_off_time: number;
  propose_depature_time: string | null;
  propose_arrival_time: string | null;
  operator_code: string;
  propose_trsp_ability: ProposeTrspAbility[] | null;
  request_snapshot?: RequestSnapshot;
}

export interface ProposeTrspAbility {
  id: number;
  day: string;
  created_at: string;
  created_date: string;
  giai: string;
  price: number;
  status: number;
  trip_name: string;
  arrival_to: number;
  assign_type: null;
  car_info_id: number;
  max_payload: null;
  operator_id: string;
  total_width: number;
  arrival_time: string;
  total_height: number;
  total_length: number;
  vehicle_size: null;
  vehicle_type: null;
  display_order: null;
  operator_code: string;
  departure_from: number;
  adjustment_price: null;
  vehicle_avb_resource_id: number;
  car_ctrl_num_id: number;
  tractor_idcr: number;
  trailer_license_plt_num_id: number;
  departure_time_min: string;
  departure_time_max: string;
  car_info: CarInfo[] | null;
}

export interface CarInfo {
  id: number;
  giai: string;
  service_no: string;
  pcke_frm_cd: null;
  car_wid_meas: number;
  freight_rate: number;
  service_name: string;
  tractor_idcr: null;
  car_hght_meas: number;
  car_lngh_meas: number;
  car_weig_meas: null;
  car_body_num_cd: null;
  car_ctrl_num_id: null;
  pcke_frm_name_cd: null;
  service_end_date: null;
  service_end_time: null;
  car_cls_of_shp_cd: null;
  service_strt_date: string;
  service_strt_time: string;
  car_cls_of_size_cd: null;
  car_max_untl_cp_quan: null;
  car_trk_bed_wid_meas: null;
  trms_of_lwr_tmp_meas: null;
  trms_of_upp_tmp_meas: null;
  car_max_load_vol_meas: null;
  car_trk_bed_hght_meas: null;
  car_trk_bed_lngh_meas: null;
  car_cls_of_crn_exst_cd: null;
  car_cls_of_rfg_exst_cd: null;
  car_license_plt_num_id: null;
  car_rmk_about_eqpm_txt: null;
  car_vol_of_hzd_item_meas: number | null;
  car_trk_bed_grnd_hght_meas: number | null;
  trailer_license_plt_num_id: string | null;
  car_cls_of_tlg_lftr_exst_cd: string | null;
  car_max_load_capacity1_meas: number;
  car_max_load_capacity2_meas: number | null;
  car_cls_of_wing_body_exst_cd: number | null;
  car_spc_grv_of_hzd_item_meas: number | null;
  car_cmpn_name_of_gtp_crtf_exst_txt: string | null;
  vehicle_name: string | null;
  temperature_range: string[] | null;
  display_order: number | null;
  cut_off_fee: number;
  cut_off_time: number;
  order_status: number;
}